packages/sdk/hush-py/pyproject.toml
[project]
name = "clawdstrike"
version = "0.2.0"
description = "Python SDK for clawdstrike security verification"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
authors = [
    { name = "Clawdstrike Contributors" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Typing :: Typed",
]
dependencies = [
    "pynacl>=1.5.0",
    "pyyaml>=6.0",
    "pycryptodome>=3.19.0",
    "wcmatch>=8.5",
]

[project.optional-dependencies]
attestation = ["httpx>=0.25"]
hunt = ["nats-py>=2.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1",
    "types-PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/clawdstrike", "src/hush"]

[tool.hatch.build.targets.wheel.force-include]
"../../../rulesets" = "clawdstrike/rulesets"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
